+++ src/kakeibo/web.clj
@@ -1,7 +1,0 @@
-(ns kakeibo.web
-  (:gen-class))
-
-(defn -main
-  "I don't do a whole lot ... yet."
-  [& args]
-  (println "Hello, Kakeibo!"))